consolidate `findFile`/`findNearest`
